requirements.txt
groq
streamlit
fastapi
uvicorn
pdfplumber
python-docx
requests
pytest
pytest-cov
python-dotenv
pydantic
httpx